fix QR-Code
we want to save just the filename in the DB to decouple it from the way it currently works so when we adapt it won't break (as easily) this also required some fiddling with migrations so be aware that if you jump to the commit before this one (e625a6f8f9d34c2bcb460fd412390bd19a5cf136) and apply the migration there (e.g. just run the http since we auto apply migrations) you might have some issues with the db data content in the SwitchBoxIdLinks table and will have to manually delete the wrong data via pgadmin4.